Refactor Parsing Logic & Add Unit Tests
This will merely refactor all the parsing-related content into a new
parser module. The pipeline parsing logic was removed from the peg
module and placed into it's own module. Unit tests were added for the
new `for` grammar and statement splitter.
