+++ src/flow_control.rs
@@ -1,8 +1,4 @@
-use super::peg::Pipeline;
-use directory_stack::DirectoryStack;
-use variables::Variables;
-use shell_expand::words::{WordIterator, WordToken};
-use shell_expand::{braces, variables};
+use parser::peg::Pipeline;
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum Statement {
@@ -76,81 +72,4 @@ impl Default for FlowControl {
             current_statement: Statement::Default,
         }
     }
-}
-
-pub enum ForKind {
-    Normal(String),
-    Range(usize, usize)
-}
-
-pub fn parse_for(expression: &str, dir_stack: &DirectoryStack, variables: &Variables) -> ForKind {
-    let mut output = String::new();
-    let mut word_iterator = WordIterator::new(expression);
-
-    let expand_variable = |variable: &str, _: bool| {
-        variables.get_var(variable)
-    };
-    let expand_command = |command:  &str, quoted: bool| {
-        variables.command_expansion(command, quoted)
-    };
-
-    while let Some(Ok(word)) = word_iterator.next() {
-        match word {
-            WordToken::Brace(text, contains_variables) => {
-                if contains_variables {
-                    let mut temp = String::new();
-                    variables::expand(&mut temp, text,
-                        |variable| expand_variable(variable, false),
-                        |command| expand_command(command, false)
-                    );
-                    braces::expand_braces(&mut output, &temp);
-                } else {
-                    braces::expand_braces(&mut output, text);
-                }
-            },
-            WordToken::Normal(expr) => output.push_str(expr),
-            WordToken::Tilde(tilde) => match variables.tilde_expansion(tilde, dir_stack) {
-                Some(expanded) => output.push_str(&expanded),
-                None           => output.push_str(tilde),
-            },
-            WordToken::Variable(text, quoted) => {
-                variables::expand(&mut output, text,
-                    |variable| expand_variable(variable, quoted),
-                    |command| expand_command(command, quoted)
-                );
-            }
-        }
-    }
-
-    {
-        let mut bytes_iterator = output.bytes().enumerate();
-        while let Some((id, byte)) = bytes_iterator.next() {
-            match byte {
-                b'0'...b'9' => continue,
-                b'.' => match output[0..id].parse::<usize>().ok() {
-                    Some(first_number) => {
-                        let mut dots = 1;
-                        for (_, byte) in bytes_iterator {
-                            if byte == b'.' { dots += 1 } else { break }
-                        }
-
-                        match output[id+dots..].parse::<usize>().ok() {
-                            Some(second_number) => {
-                                match dots {
-                                    2 => return ForKind::Range(first_number, second_number),
-                                    3 => return ForKind::Range(first_number, second_number+1),
-                                    _ => break
-                                }
-                            },
-                            None => break
-                        }
-                    },
-                    None => break
-                },
-                _ => break
-            }
-        }
-    }
-
-    ForKind::Normal(output)
 }
