+++ src/parser/mod.rs
@@ -1,0 +1,8 @@
+mod for_expression;
+pub mod peg;
+pub mod pipelines;
+pub mod shell_expand;
+mod statements;
+
+pub use self::for_expression::ForExpression;
+pub use self::statements::StatementSplitter;
